pyproject.toml
[tool.poetry]
name = "open-earn"
version = "0.1.0"
description = ""
authors = ["Morgan C. Nicholson <nich.morgan@outlook.com>"]
readme = "README.md"
packages = [{include = "open_earn"}]

[tool.poetry.dependencies]
python = "^3.9"
jinja2 = "^3.1.2"
uvicorn = "^0.20.0"
fastapi = {version = "^0.92.0", extras = ["all"]}
pydantic = "^1.10.5"
antidote = "^2.0.0"
gitpython = "^3.1.31"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
